+++ src/Facades/FeatureAccess.php
@@ -20,10 +20,14 @@ class FeatureAccess extends Facade
         $feature = config('feature-access.'.$feature_name);
         unset($feature['levels']);
 
+
+
         if (! is_null($level)) {
             $feature_level = config('feature-access.'.$feature_name.'.levels.'.$level);
-            $feature = array_merge($feature, $feature_level);
-            $feature['level'] = $level;
+            if (is_array($feature_level)){
+                $feature = array_merge($feature, $feature_level);
+                $feature['level'] = $level;
+            }
         }
 
         return $feature;